+++ Example/Sources/MockMessage.swift
@@ -58,40 +58,40 @@ struct MockMessage: MessageType {
     var messageId: String
     var sender: Sender
     var sentDate: Date
-    var data: MessageData
+    var kind: MessageKind
 
-    private init(data: MessageData, sender: Sender, messageId: String, date: Date) {
-        self.data = data
+    private init(kind: MessageKind, sender: Sender, messageId: String, date: Date) {
+        self.kind = kind
         self.sender = sender
         self.messageId = messageId
         self.sentDate = date
     }
 
     init(text: String, sender: Sender, messageId: String, date: Date) {
-        self.init(data: .text(text), sender: sender, messageId: messageId, date: date)
+        self.init(kind: .text(text), sender: sender, messageId: messageId, date: date)
     }
 
     init(attributedText: NSAttributedString, sender: Sender, messageId: String, date: Date) {
-        self.init(data: .attributedText(attributedText), sender: sender, messageId: messageId, date: date)
+        self.init(kind: .attributedText(attributedText), sender: sender, messageId: messageId, date: date)
     }
 
     init(image: UIImage, sender: Sender, messageId: String, date: Date) {
         let mediaItem = MockMediaItem(image: image)
-        self.init(data: .photo(mediaItem), sender: sender, messageId: messageId, date: date)
+        self.init(kind: .photo(mediaItem), sender: sender, messageId: messageId, date: date)
     }
 
     init(thumbnail: UIImage, sender: Sender, messageId: String, date: Date) {
         let mediaItem = MockMediaItem(image: thumbnail)
-        self.init(data: .video(mediaItem), sender: sender, messageId: messageId, date: date)
+        self.init(kind: .video(mediaItem), sender: sender, messageId: messageId, date: date)
     }
 
     init(location: CLLocation, sender: Sender, messageId: String, date: Date) {
         let locationItem = MockLocationItem(location: location)
-        self.init(data: .location(locationItem), sender: sender, messageId: messageId, date: date)
+        self.init(kind: .location(locationItem), sender: sender, messageId: messageId, date: date)
     }
 
     init(emoji: String, sender: Sender, messageId: String, date: Date) {
-        self.init(data: .emoji(emoji), sender: sender, messageId: messageId, date: date)
+        self.init(kind: .emoji(emoji), sender: sender, messageId: messageId, date: date)
     }
 
 }