+++ Sources/Views/Cells/LocationMessageCell.swift
@@ -64,7 +64,7 @@ open class LocationMessageCell: MessageContentCell {
         let annotationView = displayDelegate.annotationViewForLocation(message: message, at: indexPath, in: messagesCollectionView)
         let animationBlock = displayDelegate.animationBlockForLocation(message: message, at: indexPath, in: messagesCollectionView)
 
-        guard case let .location(locationItem) = message.data else { fatalError("") }
+        guard case let .location(locationItem) = message.kind else { fatalError("") }
 
         activityIndicator.startAnimating()